% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{check_lambda}
\alias{check_lambda}
\title{check lambda}
\usage{
check_lambda(lambda, infb, supb)
}
\arguments{
\item{lambda}{Numeric, value of lambda.}

\item{infb}{Numeric, lower bound of lambda.}

\item{supb}{Numeric, upper bound of lambda.}
}
\value{
lambda Numeric, valid value of lambda.
}
\description{
check lambda
}
