% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{choice_p}
\alias{choice_p}
\title{choice model}
\usage{
choice_p(effect)
}
\arguments{
\item{effect}{Factor, simple, fixed or lasso.}
}
\value{
penalty Numeric, 1, 2 and 3.
}
\description{
choice model
}
