% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_psi_als}
\alias{d_psi_als}
\title{D Psi ALS}
\usage{
d_psi_als(x, tau)
}
\arguments{
\item{x}{generic vector}

\item{tau}{percentile}
}
\value{
y vector, linear transformation by derivative ALS psi
}
\description{
Derivative of Psi asymetric least square
}
