% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{d_psi_mq}
\alias{d_psi_mq}
\title{D Psi M-quantile}
\usage{
d_psi_mq(x, tau, c)
}
\arguments{
\item{x}{generic vector}

\item{tau}{percentile}

\item{c}{tuning}
}
\value{
y vector, linear transformation by second derivative m-rho
}
\description{
Derivative of psi M-quantile
}
