% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{f_den}
\alias{f_den}
\title{Kernel density}
\usage{
f_den(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
y vector, kernel density estimation.
}
\description{
Kernel density
}
\examples{
x = rnorm(10)
f_den(x)

}
