% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{f_tab}
\alias{f_tab}
\title{Tabular function}
\usage{
f_tab(N, n, d, theta, sig2, kind)
}
\arguments{
\item{N}{sample size.}

\item{n}{length of alpha.}

\item{d}{length of beta.}

\item{theta}{Numeric vector.}

\item{sig2}{Numeric vector.}

\item{kind}{Numeric, 1 means alpha, 2 means beta}
}
\value{
a list with a dataframe Core and a matrix Matx, both display the same information
}
\description{
Tabular function
}
