% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_mqr}
\alias{loss_mqr}
\title{Loss M-quantile regression}
\usage{
loss_mqr(beta, x, y, tau, N, d, c)
}
\arguments{
\item{beta}{initial values}

\item{x}{design matrix}

\item{y}{vector output}

\item{tau}{percentile}

\item{N}{sample size}

\item{d}{columns of x}

\item{c}{tuning}
}
\value{
eta Numeric, sum of M-quantile regression
}
\description{
This function returns the core of M-quantile regression to be minimized
}
