% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_mqrfe}
\alias{loss_mqrfe}
\title{Loss M-quantile regression with fixed effects}
\usage{
loss_mqrfe(theta, x, y, z, tau, n, d, mm, c)
}
\arguments{
\item{theta}{initial values}

\item{x}{design matrix}

\item{y}{vector output}

\item{z}{incident matrix}

\item{tau}{percentile}

\item{n}{N sample size}

\item{d}{columns of x}

\item{mm}{n columns of z}

\item{c}{tuning}
}
\value{
eta Numeric, sum of M-quantile regression with fixed effects
}
\description{
This function returns the core of M-quantile regression with fixed effects to be minimized
}
