% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loss_mqrlasso}
\alias{loss_mqrlasso}
\title{Loss lasso M-quantile regression with fixed effects}
\usage{
loss_mqrlasso(theta, x, y, z, tau, n, d, mm, c, lambda)
}
\arguments{
\item{theta}{initial values}

\item{x}{design matrix}

\item{y}{vector output}

\item{z}{incident matrix}

\item{tau}{percentile}

\item{n}{N sample size}

\item{d}{columns of x}

\item{mm}{n columns of z}

\item{c}{tuning}

\item{lambda}{constriction parameter}
}
\value{
eta Numeric, sum of lasso M-quantile regression with fixed effects
}
\description{
This function returns the core of lasso M-quantile regression with fixed effects to be minimized
}
