% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{optim_er}
\alias{optim_er}
\title{optim expectile regression}
\usage{
optim_er(beta, x, y, tau, N, d)
}
\arguments{
\item{beta}{Numeric vector, initials values beta.}

\item{x}{Numeric matrix, covariates.}

\item{y}{Numeric vector, output.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{d}{Numeric integer, X number of columns.}
}
\value{
parametric vector and residuals.
}
\description{
This function solves a expectile regression
}
