% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{optim_erfe}
\alias{optim_erfe}
\title{optim expectile regression with fixed effects}
\usage{
optim_erfe(beta, alpha, x, y, z, tau, N, d, n)
}
\arguments{
\item{beta}{Numeric vector, initials values beta.}

\item{alpha}{Numeric vector, initials values alpha.}

\item{x}{Numeric matrix, covariates.}

\item{y}{Numeric vector, output.}

\item{z}{Numeric matrix, incidence matrix.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{d}{Numeric integer, X number of columns.}

\item{n}{Numeric integer, length of alpha.}
}
\value{
parametric vector and residuals.
}
\description{
This function solves a expectile regression  with fixed effects
}
