% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{optim_mqr}
\alias{optim_mqr}
\title{optim M-quantile regression}
\usage{
optim_mqr(beta, x, y, tau, N, d, c)
}
\arguments{
\item{beta}{Numeric vector, initials values beta.}

\item{x}{Numeric matrix, covariates.}

\item{y}{Numeric vector, output.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{d}{Numeric integer, X number of columns.}

\item{c}{Numeric, positive real value.}
}
\value{
parametric vector and residuals.
}
\description{
This function solves a M-quantile regression
}
