% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{optim_mqrfe}
\alias{optim_mqrfe}
\title{optim quantile regression with fixed effects}
\usage{
optim_mqrfe(beta, alpha, x, y, z, tau, N, d, n, c)
}
\arguments{
\item{beta}{Numeric vector, initials values beta.}

\item{alpha}{Numeric vector, initials values alpha.}

\item{x}{Numeric matrix, covariates.}

\item{y}{Numeric vector, output.}

\item{z}{Numeric matrix, incidence matrix.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{d}{Numeric integer, X number of columns.}

\item{n}{Numeric integer, length of alpha.}

\item{c}{Numeric, positive real value.}
}
\value{
parametric vector and residuals.
}
\description{
This function solves a quantile regression  with fixed effects
}
