% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{optim_qr}
\alias{optim_qr}
\title{optim quantile regression}
\usage{
optim_qr(beta, x, y, tau, N, d)
}
\arguments{
\item{beta}{Numeric vector, initials values.}

\item{x}{Numeric matrix, covariates.}

\item{y}{Numeric vector, output.}

\item{tau}{Numeric scalar, the percentile.}

\item{N}{Numeric integer, sample size.}

\item{d}{Numeric integer, X number of columns.}
}
\value{
parametric vector and residuals.
}
\description{
This function solves a quantile regression
}
