% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{print.PQR}
\alias{print.PQR}
\title{Print an PQR}
\usage{
\method{print}{PQR}(x, ...)
}
\arguments{
\item{x}{An object of class "PQR"}

\item{...}{further arguments passed to or from other methods.}
}
\value{
None
}
\description{
Define the visible part of the object class PQR
}
