% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_als}
\alias{psi_als}
\title{Psi ALS}
\usage{
psi_als(x, tau)
}
\arguments{
\item{x}{generic vector}

\item{tau}{percentile}
}
\value{
y vector, linear transformation by ALS psi
}
\description{
Psi asymetric least square
}
