% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psi_mq}
\alias{psi_mq}
\title{Psi M-quantile}
\usage{
psi_mq(x, tau, c)
}
\arguments{
\item{x}{generic vector}

\item{tau}{percentile}

\item{c}{tuning}
}
\value{
y vector, linear transformation by m-rho derivative
}
\description{
Psi M-quantile
}
