% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{q_cov}
\alias{q_cov}
\title{Covariance}
\usage{
q_cov(n, N, d, Z, X, tau, res, penalty, c)
}
\arguments{
\item{n}{length of alpha.}

\item{N}{sample size.}

\item{d}{length of beta.}

\item{Z}{Numeric matrix, incident matrix.}

\item{X}{Numeric matrix, covariates.}

\item{tau}{Numeric, identifies the percentile.}

\item{res}{Numeric vector, residuals.}

\item{penalty}{Numeric, 1 quantile regression, 2 quantile regression with fixed effects, 3 Lasso quantile regression with fixed effects}

\item{c}{Numeric, tuning}
}
\value{
a list with two matrices: sig2_alpha (which is the matrix of covariance of estimated alpha) and sig2_beta (which is the matrix of covariance of estimated beta)
}
\description{
Estimate Covariance matrix
}
