% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_koenker}
\alias{rho_koenker}
\title{Rho Koenker}
\usage{
rho_koenker(x, tau)
}
\arguments{
\item{x}{generic vector}

\item{tau}{percentile}
}
\value{
y vector, linear transformation by rho
}
\description{
Rho Koenker
}
