% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rho_mq}
\alias{rho_mq}
\title{Rho M-quantile}
\usage{
rho_mq(x, tau, c)
}
\arguments{
\item{x}{generic vector}

\item{tau}{percentile}

\item{c}{tuning}
}
\value{
y vector, linear transformation by m-rho
}
\description{
Rho M-quantile
}
