% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PQRFE_main.R
\name{sgf}
\alias{sgf}
\title{Identify significance}
\usage{
sgf(x)
}
\arguments{
\item{x}{Numeric vector.}
}
\value{
y vector Factor, symbol flag of significant p-values.

a vector of Factors, i.e., the symbols to help p-value interpretation
}
\description{
Identify significance
}
\examples{
n = 10
pvalue = rgamma(10,1,10)
sgf(pvalue) 

}
